+++ app/controller/devices.py
@@ -1,36 +1,34 @@
-# app/routes/history.py
-import hashlib
+# app/controller/devices.py
 from flask import Blueprint, request, jsonify
 from flask_cors import cross_origin
 from app.services.device_service import *
 from app.utils.influx import purge
 
-'''
-    Ce fichier représente une route serveur.
-'''
-
 devices_bp = Blueprint('devices', __name__, url_prefix = '/devices')
+service = DeviceService()
 
 @devices_bp.route('/', methods = ['GET'])
 @cross_origin()
 def get_devices():
     # Appelle le service pour renvoyer les appareils.
-    data = recuperer_appareils()
+    data = service.recuperer_appareils()
 
     # Renvoie un code 204 en cas d'absence de données.
     if not data: return '', 204
 
+    # Renvoie une réponse 200 (OK) ainsi que la liste des capteurs sous forme de json.
     return jsonify(data), 200
 
 @devices_bp.route('/<int:id>', methods = ['GET'])
 @cross_origin()
 def recuperer_capteur(id: int):
     # Appelle le service pour renvoyer les appareils.
-    capteur = recuperer_appareil(id)
+    capteur = service.recuperer_appareil(id)
 
     # Renvoie un code 204 en cas d'absence de données.
     if not capteur: return '', 204
 
+    # Renvoie une réponse 200 (OK) ainsi que la liste des capteurs sous forme de json.
     return jsonify(capteur), 200
 
 @devices_bp.route('/<string:input>/find', methods = ['GET'])
@@ -43,7 +41,7 @@ def find_devices_by_input(input: str):
     if not input or len(input) < 3: return '', 204
 
     # Appelle le service pour ajouter le capteur (data correspond au modèle device).
-    appareils = recuperer_appareils_par_saisie(input)
+    appareils = service.recuperer_appareils_par_saisie(input)
 
     # Renvoie un statut code (créé) et le capteur ajouté (il faudrait avoir un callback côté service car là on reprend ce qui a été envoyé).
     return jsonify(appareils), 200
@@ -66,22 +64,25 @@ def ajouter_device():
         'comment': capteur.get('comment', None)
     }
 
-    # On vire les espaces blancs en début et fin de chaque entrée si c'est une chaîne.
-    trimmed = {key: value.strip() if isinstance(value, str) else value for key, value in capteur_validation.items()}
+    # On vire les espaces blancs en début et fin de chaque données s'il s'agit une chaîne avec strip().
+    # Ça se lit comme ça :
+    # Pour chaque paire clé-valeur dans capteur_validation, si la valeur est une chaîne de caractères, alors retire les espaces superflus.
+    # Sinon, conserve la valeur telle quelle et assigne le résultat à la nouvelle paire clé-valeur `trimmed` (qui est lui aussi un dictionnaire).
+    capteur_propre = {key: value.strip() if isinstance(value, str) else value for key, value in capteur_validation.items()}
 
     # On doit s'assurer qu'il y a bien quelque chose dans le corps de la requête (comme le nom).
     if not capteur or 'tag' not in capteur: return jsonify(capteur), 400
 
     # Vérifie que le nom n'est pas déjà utilisé par un autre capteur existant.
-    capteurExistant = recuperer_appareil_par_nom(capteur['tag'])
+    capteurExistant = service.recuperer_appareil_par_tag(capteur['tag'])
 
     # Si le capteur existe déjà.
     if capteurExistant: return jsonify(f"Ce tag est déjà utilisé par le capteur `{capteurExistant['nameGtc']}`."), 400
 
     # Appelle le service pour ajouter le capteur (data correspond au modèle device).
-    ajouter_appareil(trimmed)
+    service.ajouter_appareil(capteur_propre)
 
-    # Renvoie un statut code (créé) et le capteur ajouté (il faudrait avoir un callback côté service car là on reprend ce qui a été envoyé).
+    # Renvoie un statut code 201 (créé) et le capteur ajouté.
     return jsonify(capteur), 201
 
 @devices_bp.route('/<int:id>/update', methods = ['PUT'])
@@ -93,9 +94,9 @@ def modifier_device(id: int):
     if not capteur or 'id' not in capteur: return jsonify(capteur), 400
 
     # Appelle le service pour ajouter le capteur (data correspond au modèle device).
-    modifier_appareil(id, capteur)
+    service.modifier_appareil(id, capteur)
 
-    # Renvoie un statut code (créé) et le capteur ajouté (il faudrait avoir un callback côté service car là on reprend ce qui a été envoyé).
+    # Renvoie un statut code (créé) et le capteur ajouté.
     return jsonify(capteur), 200
 
 @devices_bp.route('/<int:id>/<string:tag>/delete', methods = ['DELETE'])
@@ -105,7 +106,7 @@ def supprimer_device(id: int, tag: str):
     purge(tag)
     
     # Supprime l'appareil.
-    supprimer_appareil(id)
+    service.supprimer_appareil(id)
 
-    # Renvoie un statut code (créé) et le capteur ajouté (il faudrait avoir un callback côté service car là on reprend ce qui a été envoyé).
+    # Renvoie un statut code (créé) et le capteur ajouté.
     return jsonify('Appareil supprimé avec succès.'), 200
